crimereport/requirements.txt
django~=3.1.1
djangorestframework==3.12.1
mysqlclient==2.0.1
pymysql==0.10.1
defusedxml==0.6.0 
django-allauth==0.43.0 
oauthlib==3.1.0 
pyjwt==1.7.1 
python3-openid==3.2.0 
requests-oauthlib==1.3.0
pillow == 7.2.0